+++ core/incremental/mod.rs
@@ -2,6 +2,7 @@ pub mod compiler;
 pub mod cursor;
 pub mod dbsp;
 pub mod expr_compiler;
+pub mod input_operator;
 pub mod operator;
 pub mod persistence;
 pub mod view;
